apps/api/requirements.txt
fastapi
uvicorn
python-multipart
provider-hub
pillow
pyyaml
playwright
aiofiles
requests
numpy
opencv-python
torch
open-clip-torch
transformers
faiss-cpu
cairosvg
